+++ docs/fr_FR/Changelog.md
@@ -17,6 +17,7 @@
 - Améliorations: Remplacement modèle cmde 'setTemperatureLight' => cmd 'act_setColorTemp-K'.
 - Améliorations: Remplacement modèle cmde 'setTemperatureLight_1700_4900' => cmd 'act_setColorTemp-K'.
 - Améliorations: Remplacement modèle cmde 'setTemperatureLight_2700_6500' => cmd 'act_setColorTemp-K'.
+- Améliorations: Remplacement modèle cmde 'setTemperatureLight1/setTemperatureLight2' => cmd 'act_setColorTemp-K'.
 
 ## 251228-BETA-1